Allow named volumes for external drivers.
