+++ daemon/volumes_experimental.go
@@ -3,6 +3,9 @@
 package daemon
 
 import (
+	"path/filepath"
+
+	"github.com/docker/docker/runconfig"
 	"github.com/docker/docker/volume"
 	"github.com/docker/docker/volume/drivers"
 )
@@ -12,4 +15,12 @@ func getVolumeDriver(name string) (volume.Driver, error) {
 		name = volume.DefaultDriverName
 	}
 	return volumedrivers.Lookup(name)
+}
+
+func parseVolumeSource(spec string, config *runconfig.Config) (string, string, error) {
+	if !filepath.IsAbs(spec) {
+		return spec, "", nil
+	}
+
+	return "", spec, nil
 }
