+++ daemon/volumes_stubs_unit_test.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"testing"
 
+	"github.com/docker/docker/runconfig"
 	"github.com/docker/docker/volume"
 	"github.com/docker/docker/volume/drivers"
 	"github.com/docker/docker/volume/local"
@@ -31,5 +32,50 @@ func TestGetVolumeDefaultDriver(t *testing.T) {
 
 	if d.Name() != volume.DefaultDriverName {
 		t.Fatalf("Expected local driver, was %s\n", d.Name)
+	}
+}
+
+func TestParseBindMount(t *testing.T) {
+	cases := []struct {
+		bind      string
+		expDest   string
+		expSource string
+		expName   string
+		expRW     bool
+		fail      bool
+	}{
+		{"/tmp:/tmp", "/tmp", "/tmp", "", true, false},
+		{"/tmp:/tmp:ro", "/tmp", "/tmp", "", false, false},
+		{"/tmp:/tmp:rw", "/tmp", "/tmp", "", true, false},
+		{"/tmp:/tmp:foo", "/tmp", "/tmp", "", false, true},
+		{"name:/tmp", "", "", "", false, true},
+		{"local/name:/tmp:rw", "", "", "", true, true},
+	}
+
+	for _, c := range cases {
+		conf := &runconfig.Config{}
+		m, err := parseBindMount(c.bind, conf)
+		if c.fail {
+			if err == nil {
+				t.Fatalf("Expected error, was nil, for spec %s\n", c.bind)
+			}
+			continue
+		}
+
+		if m.Destination != c.expDest {
+			t.Fatalf("Expected destination %s, was %s, for spec %s\n", c.expDest, m.Destination, c.bind)
+		}
+
+		if m.Source != c.expSource {
+			t.Fatalf("Expected source %s, was %s, for spec %s\n", c.expSource, m.Source, c.bind)
+		}
+
+		if m.Name != c.expName {
+			t.Fatalf("Expected name %s, was %s for spec %s\n", c.expName, m.Name, c.bind)
+		}
+
+		if m.RW != c.expRW {
+			t.Fatalf("Expected RW %v, was %v for spec %s\n", c.expRW, m.RW, c.bind)
+		}
 	}
 }
